Fix incorrect config file in designate-api
Removes a straggler from the "one-conf-to-rule-them-all" approach used
previously that was rendering a part of the API config useless.  This
was resulting in an error where the API couldn't find keystone.

Also removes a reference to memcached and a cut-and-paste error from
some other operator to add another volume for sharing data with a
pod-mate container.